Consolidate `contract-activity.tsx` into `contract-tabs.tsx`
